Remove all forms stuff
