--- .env.example
@@ -1,3 +1,0 @@
-SENDGRID_URL=
-SENDGRID_KEY=
-ERROR_EMAIL=
