+++ chrome/manifest.json
@@ -19,5 +19,5 @@
     "16": "script-icon_16.png",
     "128": "script-icon_128.png"
   },
-  "version": "8.0.0"
+  "version": "8.0.1"
 }
