+++ src/interfaces/html-templates.ts
@@ -1,22 +1,5 @@
 /**
  * Interface for Settings Template.
- * @interface SettingsTemplate
- * @extends {GoogleAppsScript.HTML.HtmlTemplate}
- */
-interface IFormsSettingsTemplate extends GoogleAppsScript.HTML.HtmlTemplate {
-  [key: string]: string | any;
-  apiKey: string;
-  apiUrl: string;
-  apiToken: string;
-  activityId: string;
-  text1: string;
-  int1: string;
-  int2: string;
-  date1: string;
-}
-
-/**
- * Interface for Settings Template.
  * @interface SheetsSettingsTemplate
  * @extends {FormsSettingsTemplate}
  */
@@ -37,13 +20,4 @@ interface ISheetsSettingsTemplate extends GoogleAppsScript.HTML.HtmlTemplate {
   lastName: string;
   verified: string;
   issued: string;
-}
-
-/**
- * Interface for Auth Template.
- * @interface AuthTemplate
- * @extends {GoogleAppsScript.HTML.HtmlTemplate}
- */
-interface IAuthTemplate extends GoogleAppsScript.HTML.HtmlTemplate {
-  authUrl: string;
 }
