--- src/templates/auth.email.html
@@ -1,22 +1,0 @@
-<p>
-  The Google Forms add-on
-  <i>OpenBadges</i> is set to run automatically when a form is submitted. The add-on was recently updated and it needs you to
-  re-authorize it to run on your behalf.
-</p>
-
-<p>
-  The add-on's automatic functions are temporarily disabled until you re-authorize the add-on. You can accomplish this by opening
-  one of the forms using the add-on and running the add-on through the menu. Alternatively, you can click this link to authorize
-  the add-on directly:
-</p>
-
-<p>
-  <a href="<?= authUrl ?>">Click here</a> to re-authorize the add-on.
-</p>
-
-<hr>
-
-<p style="font-size:80%">
-  This automatic message was sent to you via the
-  <i>OpenBadges</i> add-on for Google Forms.
-</p>
