--- src/templates/auth.modal.html
@@ -1,23 +1,0 @@
-<link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/materialize/0.100.2/css/materialize.min.css">
-<link rel="stylesheet" href="https://fonts.googleapis.com/css?family=Roboto:300,400,500,700" type="text/css">
-<style>
-  .collection-item {
-    padding: 10px 20px;
-  }
-</style>
-<main>
-  <form id="configuration-form">
-    <ul class="collection" style="border: 0; margin: 0">
-      <li class="collection-item">
-        <div>
-          This add-on needs you to authorize it so that it can run. This is sometimes required if the add-on has updated and is using Google APIs which it was not using before.
-        </div>
-      </li>
-      <li class="collection-item">
-        <div class="col s12 center-align">
-          <a href="<?= authUrl ?>" target="_blank" class="waves-effect waves-light btn blue">Authorize</a>
-        </div>
-      </li>
-    </ul>
-  </form>
-</main>
