--- src/templates/forms-settings.sidebar.html
@@ -1,88 +1,0 @@
-<link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/materialize/0.100.2/css/materialize.min.css">
-<link rel="stylesheet" href="https://fonts.googleapis.com/css?family=Roboto:300,400,500,700" type="text/css">
-
-<style>
-  .collection-item {
-    padding: 10px 20px;
-  }
-</style>
-<main>
-  <form id="configuration-form">
-    <ul class="collection" style="border: 0; margin: 0">
-      <li class="collection-item grey lighten-3">
-        <div>
-          <strong>Settings</strong>
-        </div>
-        <div>
-          Properties can be dynamically retrieved from the form response using the following format <code>{{Question title}}</code>.
-        </div>
-      </li>
-      <li class="collection-item">
-        <div class="input-field">
-          <input id="apiUrl" name="apiUrl" type="url" class="validate" required value="<?= apiUrl ?>">
-          <label for="apiUrl" data-error="Not a valid URL">API URL</label>
-        </div>
-        <div class="input-field">
-          <input id="apiKey" name="apiKey" type="text" class="validate" required value="<?= apiKey ?>">
-          <label for="apiKey">API Key</label>
-        </div>
-        <div class="input-field">
-          <input id="apiToken" name="apiToken" type="text" class="validate" required value="<?= apiToken ?>">
-          <label for="apiToken">API Token</label>
-        </div>
-        <div class="input-field">
-          <input id="activityId" name="activityId" type="text" class="validate" required value="<?= activityId ?>">
-          <label for="activityId">Activity ID</label>
-        </div>
-      </li>
-      <li class="collection-item grey lighten-3">
-        <div>
-          <strong>Extra properties</strong>
-        </div>
-        <div>
-          These are optional properties which can be sent with the request when a form is completed.
-        </div>
-      </li>
-      <li class="collection-item">
-        <div class="input-field">
-          <input id="text1" name="text1" type="text" data-length="255" class="validate" value="<?= text1 ?>">
-          <label for="text1">Text</label>
-        </div>
-        <div class="input-field">
-          <input id="int1" name="int1" type="number" step="1" class="validate" value="<?= int1 ?>">
-          <label for="int1">Integer 1</label>
-        </div>
-        <div class="input-field">
-          <input id="int2" name="int2" type="number" step="1" class="validate" value="<?= int2 ?>">
-          <label for="int2">Integer 2</label>
-        </div>
-        <!-- <div class="input-field">
-          <input id="date1" name="date1" type="text" class="validate" value="<?= date1 ?>">
-          <label for="date1">Date</label>
-        </div> -->
-      </li>
-      <li class="collection-item grey lighten-3">
-        <div class="col s12 center-align">
-          <button type="submit" class="waves-effect waves-light btn blue">Save</button>
-        </div>
-      </li>
-    </ul>
-  </form>
-</main>
-
-<script type="text/javascript" src="https://code.jquery.com/jquery-3.2.1.min.js"></script>
-<script type="text/javascript" src="https://cdnjs.cloudflare.com/ajax/libs/materialize/0.100.2/js/materialize.min.js"></script>
-<script>
-  (function () {
-    var form = $("#configuration-form");
-    form.on("submit", function () {
-      var config = form.serializeArray().reduce(function (obj, item) {
-        obj[item.name] = item.value;
-        return obj;
-      }, {});
-      google.script.run
-        .withSuccessHandler(function () { google.script.host.close() })
-        .onSaveConfiguration(config);
-    });
-  })();
-</script>
